Abstract vprint; vprint the interaction as matched
