+++ tests/ledger/vprint.py
@@ -1,0 +1,1 @@
+../network_tests/vprint.py
