+++ tests/network_tests/vprint.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+
+
+verbose = False
+
+
+def vprint(*args, timestamp=True, **kwargs):
+    global verbose
+    if verbose:
+        if timestamp:
+            print(datetime.now(), end=" ")
+        print(*args, **kwargs)
